src/backend/pyproject.toml
[project]
name = "braves-board-backend"
version = "0.1.0"
description = ""
authors = [
    { name = "Heru Purnomo" }
]
requires-python = ">=3.10"

dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.30.0",
    "sqlalchemy[asyncio]==2.0.36",
    "asyncpg==0.30.0",
    "alembic==1.14.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "authlib==1.3.2",
    "httpx==0.27.0",
    "pydantic[email]==2.9.0",
    "python-dotenv==1.0.1",
    "pydantic-settings==2.5.2",
    "redis==5.0.3",
    "types-redis==4.6.0.20241004",
    "google-cloud-storage==2.14.0",
    "itsdangerous==2.2.0",
    "python-multipart==0.0.12",
    "prometheus-client==0.20.0",
    "structlog==24.4.0"
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]